change structure models for genre and subgenre
